backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
openai-whisper==20231117
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.24.3
celery==5.3.4
redis==5.0.1
backoff==2.2.1
tenacity==8.2.3